Wolff Updates 2

+++ Code_F21_MonteCarlo/Newman_Barkema/C++/Metropolis/Metropolis_autocorrelation.cpp
@@ -1,5 +1,5 @@
 #include "Metropolis.hpp"
-#include "Writer.hpp"
+#include "../Writer.hpp"
 
 #include <iostream>
 #include <iomanip>
@@ -10,8 +10,8 @@ const int kBin = 25; /*Parametr: Change binning of temperature*/
 const int kB = 0;
 const int kJ = 1;
 
-const int Tsrt = 0;
-const int Tfin = 5;
+const double Tsrt = 0;
+const double Tfin = 5;
 
 double isTinf = false;
